% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blendML.R
\name{blendML}
\alias{blendML}
\title{Blend ML}
\usage{
blendML(data, p_FUN = glm, p_formula, p_args = list(family = gaussian),
  ml_FUN = NULL, ml_formula, ml_args = list())
}
\arguments{
\item{data}{The data-frame containing the data}

\item{p_FUN}{A function to calculate the blended parametric regression. Defaults to R's \code{glm}}

\item{p_formula}{A two-sided formula with the dependent variable and explaining variables to be included directly in the parametric regression}

\item{p_args}{= A named list of arguments to be passed to the parametric regression function (e.g list(familiy = binomial) for logistic regression)}

\item{ml_FUN}{A function to calculate the (non-parametric) ML regression. When set to \code{NULL} (the default) the will try to load and use \code{randomForest}.}

\item{ml_formula}{A one-sided formula with dependent variables to be included in the ML regression (dependent variable is table form the parametric formula)}

\item{p_args}{= A named list of arguments to be passed to the ML regression function (e.g \code{list(familiy = binomial)} for logistic regression)}
}
\value{
The function \code{blendML} returns an S3 object with class "blendML" that contains:
\item{ort}{A list containig the outputs of the orthogonalization process (a list of \code{"lm"} objects)}
\item{ML}{The optput of the non-parametric regression (by default an object of class \code{"randomForest"})}
\item{param}{The optput of the parametric regression (by default an object of class \code{"glm"})}
}
\description{
The function \code{blendM} designed to embed the score calculated by a machine learning (ML) method into a parametric regression, in particular a generalized linear model (GLM). 
The motivation is to harness the power of ML methods where a simple interpretation of the marginal effect of some variables is not mandatory, while keeping a transparent structure with simple interpretation (and direct understanding of marginal effects) for variables of interest.  \cr \cr
Since we expect possible co-linearities to exist between the variables used by the 2 methods, and since the parameters of interest are the ones included in the parametric regression, the ML regression will run on an orthogonalized version of the dependent variables (this will be done using simple linear regression and regression on the residuals) \cr \cr
}
\examples{
# Linear regression & random forests
x <- blendML(data = mtcars, p_formula = mpg ~ disp, ml_formula =  ~ cyl + hp + wt)
summary(x$param)
# Logistic regression & random forests
x <- blendML(data = mtcars, p_formula = vs ~ disp, p_args = list(family = binomial), ml_formula =  ~ cyl + hp + wt)
summary(x$param)
}
\keyword{GLM,}
\keyword{MachineLearning,}
\keyword{blending}
