% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_times.R
\name{data_times}
\alias{data_times}
\title{Get event time data}
\usage{
data_times(S)
}
\arguments{
\item{S}{vector of survival data}
}
\value{
A vector of event times
}
\description{
Get event time data
}
\keyword{internal}
