% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext_surv_est.R
\name{ext_surv_est}
\alias{ext_surv_est}
\title{Create an external survival curve based on the expert opinion}
\usage{
ext_surv_est(
  t_info,
  S_info,
  T_max,
  times_est,
  n = 70,
  nsim = 100,
  distr = "gom"
)
}
\arguments{
\item{t_info}{A vector of times for which expert opinion is elicited}

\item{S_info}{A vector of mean survival probabilities estimated by experts corresponding to timepoints in the t_pri}

\item{T_max}{The maximum survival time to be used}

\item{n}{The number of patients to construct the artificial external dataset; default 70}

\item{nsim}{The number of simulations from the distribution of the survival curves; default 100}

\item{tp}{A vector of times for which the survival curves are to be computed}
}
\description{
Create an external survival curve based on the expert opinion
}
