% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_surv.R
\name{make_surv}
\alias{make_surv}
\title{Create survival probabilities}
\usage{
make_surv(Surv, ...)
}
\arguments{
\item{Surv}{survival analysis object}

\item{...}{Additional arguments}
}
\value{
matrix
}
\description{
Create survival probabilities
}
