% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manip_plot.R
\name{manip_plot}
\alias{manip_plot}
\title{Blended survival plot with manipulate}
\usage{
manip_plot(obs_Surv, ext_Surv, blend_interv)
}
\description{
RStudio bug
need to run base R first
manipulate(plot(1:x), x = slider(5, 10))
}
