% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pw_unif_surv.R
\name{pw_unif_surv}
\alias{pw_unif_surv}
\title{Piecewise uniform survival curve}
\usage{
pw_unif_surv(p, times, epsilon = 0.1)
}
\arguments{
\item{p}{Given survival probability constraints}

\item{times}{Times at which \code{p} are defined}

\item{epsilon}{Length between generated times points}
}
\value{
Data frame of time points \code{t} and
survival probabilities \code{S}
}
\description{
Piecewise uniform survival curve
}
\examples{
p <- c(0.8, 0.3, 0.2, 0)
times <- c(10, 20, 30, 40)
surv <- pw_unif_surv(p, times)
plot(surv, type = "l")

# sample times
x <- runif(100)
purrr::map_dbl(x, ~max(surv$t[surv$S > .x]))

}
\seealso{
\code{\link{ext_surv_sim}}
}
