% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/red_snapper_sampled_data.R
\docType{data}
\name{red_snapper_sampled}
\alias{red_snapper_sampled}
\title{2016 Texas Parks and Wildlife Red Snapper Capture-Recapture Dataset}
\format{A data frame with 398 rows and 11 variables:

\describe{
\item{id}{Identificaion number of a fishing trip}
\item{angler_ps}{Number of anglers on fishing trip, as observed in the dockside intercept, a probability sample}
\item{number_caught_ps}{Number of Red Snapper fish caught as observed in the dockside intercept, a probability sample}
\item{stratum}{The stratum of the dockside intercept, either weekday or weekend}
\item{prob}{Sampling probability}
\item{sample_weight}{Sampling weight, reciprocal of sampling probability}
\item{number_caught_np}{Number of Red Snapper fish caught, as reported by captains, a non-probability sample}
\item{anglers_np}{Number of anglers on fishing trip,as reported by captains, a non-probability sample}
\item{captured_indicator}{Indicator of whether the sampled trip also self-reported}
\item{delta_catch}{Difference between number_caught_ps and number_caught_np}
\item{delta_angler}{Difference between anglers_ps and anglers_np}}}
\usage{
red_snapper_sampled
}
\description{
A dataset from a 2016 capture-recapture sampling program
by Texas Parks and Wildlife. A probability sample of
docks was selected and interviewers boarded returning charter
fishing boats and counted the number of Red Snapper fish caught and
number of anglers aboard. The captains of such boats had the opportunity
to voluntarily self-report (via a smartphone app) the number of Red Snapper
caught and number of anglers aboard. Captains who self-reported were still
eligible to be sampled in the dockside intercept. The self-reports represent
the "capture" sample and the dockside intercept represents the "recapture" sample.
If a row in \code{red_snapper_sampled} refers to a unit which is also in the capture sample, the data frame
contains the information gathered from the recapture sample. If the row refers
to a unit in the recapture sample only, those columns for recapture sample data
contain zeros.
}
\keyword{datasets}
