% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pw_unif_surv.R
\name{sample_pw_unif_surv}
\alias{sample_pw_unif_surv}
\title{Piece-wise uniform survival curve for uncertain inputs}
\usage{
sample_pw_unif_surv(p, times, n = 2, sn = 10, epsilon = 0.1)
}
\arguments{
\item{p}{Probability}

\item{times}{Event times}

\item{n}{Number of samples}

\item{sn}{Scaling factor. This is like a pseudosample size.}

\item{epsilon}{Different in times}
}
\description{
Piece-wise uniform survival curve for uncertain inputs
}
\examples{
surv <- sample_pw_unif_surv(p = c(0.8, 0.3, 0.2, 0),
                           times = c(10, 20, 30, 40), n = 100)
matplot(surv, type = "l")
surv <- sample_pw_unif_surv(p = c(0.8, 0.3, 0.2, 0),
                           times = c(10, 20, 30, 40), n = 100, sn = 100)
matplot(surv, type = "l")

}
