% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/self_reports_data.R
\docType{data}
\name{self_reports}
\alias{self_reports}
\title{Captains' Self-Reports from 2016 Texas Parks and Wildlife Red Snapper Capture-Recapture Sampling Program}
\format{A dataframe with 117 rows and 2 variables:



\describe{
\item{anglers}{Number of anglers reported by captain for fishing trip}
\item{number_kept}{Number of fish kept, as reported by captains for a fishing trip}
}}
\usage{
self_reports
}
\description{
This dataset comes from a voluntary sample of charter fishing captains. The
captains could self-report (via a smartphone app) their catch of Red Snapper fish and the number of anglers for each
fishing trip. Each reported trip could also be sampled into a probability sample of docks along the
Texas coast. Interviewers for the dockside sample boarded the boats and observed the number of Red Snapper
caught and the number of anglers onboard. The self-reports are the "capture" sample and the dockside intercept
is the "recapture" sample.
}
\keyword{datasets}
