% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_p.R
\name{t_p}
\alias{t_p}
\title{Generalized Capture-Recapture Estimator, Extension of Pollock et al's Estimator of Total}
\usage{
t_p(data, recapture_total, captured, survey_design, na_remove = T,
  capture_units)
}
\arguments{
\item{data}{A data frame, each row is an observation from the recapture sample,
containing information gathered in the recapture sample, whether or not the unit
is also in the capture sample, and sample design information.}

\item{recapture_total}{Name of variable of interest recorded in the recapture
sample, from given data frame}

\item{captured}{Name of indicator variable of the sampled unit also being in the
capture sample, from given data frame}

\item{survey_design}{A complex survey design specified with
\code{survey::svydesign()}}

\item{na_remove}{Remove NA's? Logical}

\item{capture_units}{Total number of units in the capture sample}
}
\value{
Estimate of Total and Standard Error of Estimate
  \item{total}{Estimate of total of a variable in population}
  \item{se}{Standard error of estimate of total}
}
\description{
Make an Estimate of Total Using Pollock et al's (1994) Estimator
  Generalized by Liu et al (2017) for a Complex Sample Setting. Use number of
  capture units as auxiliary information
}
\details{
This estimator is a ratio estimator defined by: \eqn{t_p = n_1
  \frac{\hat{t}_y}{\hat{n}_1}} with ratio \eqn{ t_y / n_1}. \eqn{\hat{t}_y =
  \sum{i=1}^{N}w_i  z_i  y_i} where \eqn{z_i} is a sampling indicator,
  \eqn{w_i} is the sampling weight, and \eqn{y_i} is the value of the
  variable of interest observed in the recapture sample. There are \eqn{N}
  units in the population. \eqn{n_1} is the total number of units seen in the
  capture sample and is estimated by \eqn{\hat{n}_1 = \sum{i=1}^{N}w_i z_i
  r_i} where \eqn{r_i} is an indicator of whether the unit is a member of the
  recapture sample was seen in the capture sample.
}
\examples{
  s_design <- survey::svydesign(id = ~psu,
                               strat = ~stratum,
                               prob = ~prob,
                               nest = T,
                               data = red_snapper_sampled)

  t_p(data = red_snapper_sampled,
      recapture_total = number_caught_ps,
      captured = captured_indicator,
      survey_design = s_design,
      capture_units = nrow(self_reports))


}
