% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{weightplot}
\alias{weightplot}
\title{Plots the weights for the blending procedure}
\usage{
weightplot(x, ...)
}
\arguments{
\item{x}{A blended survival curve object obtained from \code{\link[=blendsurv]{blendsurv()}}}

\item{...}{Additional arguments}
}
\value{
\pkg{ggplot2} object
}
\description{
Plots the weights for the blending procedure
}
\references{
Che, Z., Green, N., & Baio, G. (2022). Blended Survival Curves: A New Approach
to Extrapolation for Time-to-Event Outcomes from Clinical Trials in Health
Technology Assessment. Medical Decision Making, 43(3), 299-310.
\url{doi:10.1177/0272989X221134545}.
}
\seealso{
\code{\link[=blendsurv]{blendsurv()}}
}
