% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blendsurv.R
\name{blendsurv}
\alias{blendsurv}
\title{Blend survival object}
\usage{
blendsurv(
  obs_Surv,
  ext_Surv,
  blending_interval,
  beta_params = list(alpha = 3, beta = 3)
)
}
\arguments{
\item{obs_Surv}{Observed survival curve}

\item{ext_Surv}{External survival curve}

\item{blending_interval}{Maximum and minimum values}

\item{beta_params}{coefficients of a beta distribution}
}
\description{
Blend survival object
}
