% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blendsurv.R
\name{blendsurv}
\alias{blendsurv}
\title{Blended survival object}
\usage{
blendsurv(
  obs_Surv,
  ext_Surv,
  blend_interv,
  beta_params = list(alpha = 3, beta = 3),
  times_est = seq(0, 180),
  nsim = 100
)
}
\arguments{
\item{obs_Surv, ext_Surv}{Observed and external data survival curves.
These can come from \pkg{survHE}, \pkg{INLA} or \pkg{flexsurv} fits.}

\item{blend_interv}{Maximum and minimum values for the blending interval.}

\item{beta_params}{coefficients of a beta distribution}

\item{times_est}{A vector of times for which the survival curves
are to be computed}

\item{nsim}{The number of simulations from the distribution of
the survival curves; default 100}
}
\value{
List of S for observed, external and blended curves.
}
\description{
This is the main function in the \pkg{blendR} package.
Two survival curves are supplied and blended according to
the blending distribution characterised by the blending interval and
the beta distribution parameters.
}
