% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blendsurv.R
\name{blendsurv}
\alias{blendsurv}
\title{Blend survival object}
\usage{
blendsurv(
  obs_Surv,
  ext_Surv,
  blend_interval,
  beta_params = list(alpha = 3, beta = 3)
)
}
\arguments{
\item{obs_Surv}{Observed data survival curve}

\item{ext_Surv}{External survival curve}

\item{blend_interval}{Maximum and minimum values}

\item{beta_params}{coefficients of a beta distribution}

\item{tp}{A vector of times for which the survival curves are to be computed}

\item{nsim}{The number of simulations from the distribution of the survival curves; default 100}
}
\description{
Blend survival object
}
