% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_times.R
\name{data_times}
\alias{data_times}
\alias{data_times.inla}
\alias{data_times.survHE}
\alias{data_times.flexsurvreg}
\title{Get event time data}
\usage{
data_times(surv_res)

\method{data_times}{inla}(surv_res)

\method{data_times}{survHE}(surv_res)

\method{data_times}{flexsurvreg}(surv_res)
}
\description{
Get event time data
}
