% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_times.R
\name{data_times}
\alias{data_times}
\alias{data_times.inla}
\alias{data_times.survHE}
\alias{data_times.flexsurvreg}
\alias{data_times.default}
\title{Get event time data}
\usage{
data_times(S)

\method{data_times}{inla}(S)

\method{data_times}{survHE}(S)

\method{data_times}{flexsurvreg}(S)

\method{data_times}{default}(S)
}
\arguments{
\item{S}{vector of survival data}
}
\value{
A vector of event times
}
\description{
Get event time data
}
\examples{
library(survHE)

## trial data
data("TA174_FCR", package = "blendR")

## externally estimated data
data_sim <- ext_surv_sim(t_info = 144,
                         S_info = 0.05,
                         T_max = 180)

obs_Surv <- fit.models(formula = Surv(death_t, death) ~ 1,
                       data = dat_FCR,
                       distr = "exponential",
                       method = "hmc")

ext_Surv <- fit.models(formula = Surv(time, event) ~ 1,
                       data = data_sim,
                       distr = "exponential",
                       method = "hmc")

c(data_times(obs_Surv), data_times(ext_Surv))

}
\keyword{internal}
