% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_inla_pw.R
\name{fit_inla_pw}
\alias{fit_inla_pw}
\title{Fit a Piecewise Exponential Survival Model using INLA}
\usage{
fit_inla_pw(
  inla.formula = inla.surv(death_t, death) ~ -1,
  data,
  cutpoints,
  nsim = 100,
  ...
)
}
\arguments{
\item{inla.formula}{A formula specifying the survival model. The response must
be an \code{INLA::inla.surv} object (e.g., \code{inla.surv(time, event)}). The right-hand
side specifies the linear predictor. The default fits a baseline-hazard-only model.}

\item{data}{A data frame containing the time-to-event data, including the
variables named in the \code{inla.formula}.}

\item{cutpoints}{A numeric vector of cut points used to partition the time
axis into intervals for the piecewise constant hazard.}

\item{nsim}{Number of simulations from posteriors; default 100}

\item{...}{Additional arguments to be passed directly to the \code{INLA::inla}
function (e.g., \code{control.predictor}, \code{control.family}).}
}
\value{
An object of class \code{inla}, which contains the full results of the
fitted Bayesian model.
}
\description{
A convenience wrapper to fit a Bayesian piecewise exponential model (PEM) for
survival data using the \code{INLA} package. This function simplifies the process by
automatically handling the necessary data transformation from a standard
survival format to the Poisson regression format that \code{INLA} requires.
}
\details{
This function models the baseline hazard rate as a piecewise constant function.
The \code{cutpoints} define the time intervals within which the hazard is assumed
to be constant.

Internally, the function first uses \code{INLA::inla.coxph} to convert the survival
data into the long format suitable for a Poisson generalized linear model (GLM).
It then fits this model using \code{INLA::inla}, applying a random walk of order 1
(\code{rw1}) prior to the log-hazards for each interval. This prior provides a
flexible and smoothed estimate of the baseline hazard over time.
}
\examples{
\donttest{
# INLA may require configuration on your system.
# See: https://www.r-inla.org/download-install
 if (requireNamespace("INLA", quietly = TRUE)) {
  data("dat_FCR", package = "blendR")
  head(dat_FCR)

  # Fit a simple piecewise model with intervals every 5 time units
  obs_Surv <- fit_inla_pw(data = dat_FCR, cutpoints = seq(0, 180, by = 5))

  # summary(obs_Surv)
 }
}

}
\references{
Che, Z., Green, N., & Baio, G. (2022). Blended Survival Curves: A New Approach
to Extrapolation for Time-to-Event Outcomes from Clinical Trials in Health
Technology Assessment. Medical Decision Making, 43(3), 299-310.
\url{doi:10.1177/0272989X221134545}.
}
