% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_surv.R
\name{make_surv_methods}
\alias{make_surv_methods}
\alias{make_surv}
\alias{make_surv.survHE}
\alias{make_surv.flexsurvreg}
\alias{make_surv.inla}
\title{Create survival probabilities}
\usage{
make_surv(Surv, ...)

\method{make_surv}{survHE}(Surv, t, nsim = 100, ...)

\method{make_surv}{flexsurvreg}(Surv, t = NULL, nsim = 100, ...)

\method{make_surv}{inla}(Surv, t, nsim = 100, ...)
}
\arguments{
\item{Surv}{survival analysis object}

\item{...}{Additional arguments}

\item{t}{Time points; vector}

\item{nsim}{Number of simulations; integer}
}
\value{
matrix of survival probabilities
}
\description{
These function are version of the \code{make.surv} function
from \pkg{survHE}. These are needed prior to blending.
}
