% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manip_plot.R
\name{manip_plot}
\alias{manip_plot}
\title{Blended survival plot with manipulate}
\usage{
manip_plot(obs_Surv, ext_Surv, blend_interv)
}
\arguments{
\item{obs_Surv}{Observed survival}

\item{ext_Surv}{External survival}

\item{blend_interv}{Blending interval}
}
\value{
Blended survival plot
}
\description{
RStudio bug
need to run base R first
\code{manipulate(plot(1:x), x = slider(5, 10))}
}
