% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.blended}
\alias{plot.blended}
\title{Plot Blended Survival Curves}
\usage{
\method{plot}{blended}(x, alpha = c(0.1, 0.05), ...)
}
\arguments{
\item{x}{An object of class \code{blended}, typically the output of \code{\link[=blendsurv]{blendsurv()}}.}

\item{alpha}{A numeric vector of length two specifying the opacity for the
credible interval ribbons. The first value (\code{alpha[1]}) is for the blended
curve, and the second (\code{alpha[2]}) is for the trial and external curves.}

\item{...}{Additional graphical arguments passed to the plot, such as \code{xlim},
\code{xlab}, or \code{ylab}.}
}
\value{
A \code{ggplot} object representing the survival curves.
}
\description{
S3 method for plotting objects of class \code{blended}. This function generates a
\code{ggplot2} visualization that displays the mean survival curves and 95\%
credible intervals for three components:
\enumerate{
\item The curve fitted to the original trial data (\verb{Data fitting}).
\item The curve representing external information (\verb{External info}).
\item The final, combined blended curve (\verb{Blended curve}).
}
}
\examples{
\dontshow{if (rlang::is_installed("survHEhmc")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
library(survHE)

## trial data
data("TA174_FCR", package = "blendR")

## externally estimated data
data_sim <- ext_surv_sim(t_info = 144,
                         S_info = 0.05,
                         T_max = 180)

obs_Surv <- fit.models(formula = Surv(death_t, death) ~ 1,
                       data = dat_FCR,
                       distr = "exponential",
                       method = "hmc")

ext_Surv <- fit.models(formula = Surv(time, event) ~ 1,
                       data = data_sim,
                       distr = "exponential",
                       method = "hmc")

blend_interv <- list(min = 48, max = 150)
beta_params <- list(alpha = 3, beta = 3)

ble_Surv <- blendsurv(obs_Surv, ext_Surv, blend_interv, beta_params)

plot(ble_Surv)
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=blendsurv]{blendsurv()}}, \code{\link[=weightplot]{weightplot()}}
}
