% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.blended}
\alias{plot.blended}
\title{Blended survival curve based on short-term data and external information}
\usage{
\method{plot}{blended}(ble_Surv, tp = seq(0, 180))
}
\arguments{
\item{ble_Surv}{}
}
\value{
\code{ble_Sruv} invisibly
}
\description{
Blended survival curve based on short-term data and external information
}
