% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.blended}
\alias{plot.blended}
\title{Blended survival curve based on short-term data and external information}
\usage{
\method{plot}{blended}(x, alpha = c(0.1, 0.05), ...)
}
\arguments{
\item{x}{A blended survival curve object obtain from \code{\link[=blendsurv]{blendsurv()}}}

\item{alpha}{A vector specifying the opacity of ribbon for the blended curve and other curves}

\item{...}{Additional arguments}
}
\value{
\pkg{ggplot2} object
}
\description{
Blended survival curve based on short-term data and external information
}
\seealso{
\code{\link[=blendsurv]{blendsurv()}}
}
