% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pw_unif_surv.R
\name{sample_pw_unif_surv}
\alias{sample_pw_unif_surv}
\title{Piecewise uniform survival curve for uncertain inputs}
\usage{
sample_pw_unif_surv(p, times, n = 2, ...)
}
\arguments{
\item{n}{Number of samples
x <- sample_pw_unif_surv(p = c(0.8, 0.3, 0.2, 0.01),
times = c(10, 20, 30, 40))}
}
\description{
Piecewise uniform survival curve for uncertain inputs
}
