% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/self_reports_data.R
\docType{data}
\name{self_reports}
\alias{self_reports}
\title{Self-Reports}
\format{A dataframe with 117 rows and 2 variables:

\describe{
\item{anglers}{Number of anglers reported by captain for fishing trip}
\item{number_kept}{Number of fish kept, as reported by captains for a fishing trip}
}}
\usage{
self_reports
}
\keyword{datasets}
