% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_c.R
\name{t_c}
\alias{t_c}
\title{Generalized Capture-Recapture Estimator of Total for Complex Sample,
  Using Values of Total from Capture Sample as Auxiliary Information}
\usage{
t_c(recapture_total, captured_total, survey_design, na_remove = T,
  total_from_capture)
}
\arguments{
\item{recapture_total}{Variable of Total of Interest from a Recapture Sample
Observation}

\item{captured_total}{Variable of Total of Interest from a Capture Sample
Observation}

\item{survey_design}{A complex survey design specified with
\code{survey::svydesign()}}

\item{na_remove}{Remove NA's? Logical}

\item{capture_units}{Total of Variable of Interest from all Units in the
Capture Sample}
}
\value{
Estimate of Total and Standard Error of Estimate
  \item{total}{Estimate of Total of a Variable in Population}
  \item{se}{Standard Error of Estimate of Total}
}
\description{
Make an Estimate of Total Using a Generalized Capture-Recapture
  Estimator from Liu et al (2017) for a Complex Sample Setting. Uses total of
  variable of interest from the capture sample as auxiliary information
}
\details{
This estimator is a ratio estimator defined by: \eqn{t_c = t_y^*
  *\frac{\hat{t}_y}{\hat{t}_y^*}} with ratio \eqn{ t_y / n_1}. \eqn{t_y =
  \sum{i=1}^{N}w_i  z_i  y_i} where \eqn{z_i} is a sampling indicator,
  \eqn{w_i} is the sampling weight, and \eqn{y_i} is the value of the
  variable of interest observed in the recapture sample. There are \eqn{N}
  units in the population. \eqn{t_y} is estimated by \eqn{\hat{t}_y = \sum{i
  \in s_2}y_i} and \eqn{n_1} is estimated by \eqn{\hat{n}_1 = \sum{i \in
  s_2}r_i} where \eqn{r_i} is an indicator of whether the unit seen in the
  recapture sample was seen in the capture sample. \eqn{n_1} is the total
  number of units seen in the capture sample
}
