% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_p.R
\name{t_p}
\alias{t_p}
\title{Generalized Pollock et al's Estimator of Total}
\usage{
t_p(recapture_total, captured, survey_design, na_remove = T, capture_units)
}
\arguments{
\item{recapture_total}{Variable of Total of Interest from a Recapture Sample
Observation}

\item{captured}{Indicator Variable of Unit being in Capture Sample}

\item{survey_design}{A complex survey design specified with
\code{survey::svydesign()}}

\item{na_remove}{Remove NA's? Logical}

\item{capture_units}{Total Number of Units in the Capture Sample}
}
\value{
Estimate of Total and Standard Error of Estimate
  \item{total}{Estimate of Total of a Variable in Population}
  \item{se}{Standard Error of Estimate of Total}
}
\description{
Make an Estimate of Total Using Pollock et al's (1994) Estimator
Generalized by Liu et al (2017) for a Complex Sample Setting. Use number
of capture units as auxiliary information
}
\details{
This estimator is a ratio estimator defined by: \eqn{t_p = n_1 *
\frac{\hat{t}_y}{\hat{n}_1}} with ratio \eqn{ t_y / n_1}. \eqn{t_y =
\sum{i=1}^{N}w_i  z_i  y_i} where \eqn{z_i} is a sampling indicator, \eqn{w_i} is
the sampling weight, and \eqn{y_i} is the value of the variable of interest
observed in the recapture sample. There are \eqn{N} units in the population.
\eqn{t_y} is estimated by \eqn{\hat{t}_y = \sum{i \in s_2}y_i} and
\eqn{n_1} is estimated by \eqn{\hat{n}_1 = \sum{i \in s_2}r_i} where
\eqn{r_i} is an indicator of whether the unit seen in the recapture sample
was seen in the capture sample. \eqn{n_1} is the total number of units
seen in the capture sample
}
