% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_y2.R
\name{t_y2}
\alias{t_y2}
\title{Multivariate Ratio Estimator of Total for Complex Sample, from Liu et
  al (2017)}
\usage{
t_y2(delta, captured, survey_design, na_remove = T, capture_units,
  total_from_capture)
}
\arguments{
\item{delta}{For Every Unit in the Recapture Sample, Delta is the Value of
the Variable of Interest Observed in the Recapture Sample Minus the Value
Observed in the Capture Sample. \eqn{\delta = y_i - r_iy*_i} where
\eqn{y_i} is the Value of the Variable of Interest for the ith unit in the
Recapture Sample, \eqn{r_i} is an Indicator of whether the unit is a Member
of the Capture Sample, and \eqn{y*_i} is the Value of the Variable of
Interest in the Capture Sample}

\item{captured}{Indicator Variable of Unit being in Capture Sample}

\item{survey_design}{A complex survey design specified with
\code{survey::svydesign()}}

\item{na_remove}{Remove NA's? Logical}

\item{capture_units}{Total of Variable of Interest from all Units in the
Capture Sample}

\item{captured_total}{Variable of Total of Interest from a Capture Sample
Observation}
}
\value{
Estimate of Total and Standard Error of Estimate
  \item{total}{Estimate of Total of a Variable in Population}
  \item{se}{Standard Error of Estimate of Total}
}
\description{
Make an Estimate of Total Using a Weighted Combination of
  \code{\link{t_p}} and \code{\link{t_y2}} from Liu et al (2017) for a
  Complex Sample Setting.
}
\details{
This estimator is defined by: \eqn{t_y2 = t_{y*} +
  \frac{n_1}{\hat{n}_1}(\hat{t}_y - \hat{t}_{y*}) = t_{y*} +
  n_1\hat{bar{\delta}}}. This amounts to a ratio estimator of \eqn{\delta}
  over \eqn{\hat{n_1}} with auxiliary information \eqn{n_1} added to \eqn{t_{y*}}.
  \eqn{\delta = y_i - r_iy*_i}, \eqn{\hat{t}_y = \sum{i=1}^{N}w_i  z_i y_i},
  \eqn{z_i} is a sampling indicator, \eqn{w_i} is the sampling weight,
  \eqn{y_i} is the value of the variable of interest observed in the
  Recapture sample, \eqn{y*_i} is the value of the variable of interest
  observed in the Capture sample. There are \eqn{N} units in the population.
  \eqn{y*_i} is the value of the variable of interest recorded in the
  recapture sample. \eqn{t_{y*} = \sum{i =1}^{N}r_iy*_i} \eqn{r_i} is an
  indicator of whether the unit is a member of the recapture sample.
}
