% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{weightplot}
\alias{weightplot}
\title{Plots the weights for the blending procedure}
\usage{
weightplot(x, ...)
}
\arguments{
\item{x}{A blended survival curve object obtained from \code{\link[=blendsurv]{blendsurv()}}}

\item{...}{Additional arguments}
}
\value{
\pkg{ggplot2} object
}
\description{
Plots the weights for the blending procedure
}
\seealso{
\code{\link[=blendsurv]{blendsurv()}}
}
